requirements.txt
btcpay-python==1.1.0
click==7.0
Flask==1.0.2
Flask-APScheduler==1.11.0
rq-dashboard==0.3.12
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
gunicorn==19.9.0
python-dotenv==0.9.1
python-quickbooks==0.7.5
redis==2.10.6
WTForms==2.2.1
